Update gqf_cpp.h
fix return type for range()

+++ include/gqf_cpp.h
@@ -61,7 +61,7 @@ class CQF {
 		bool is_exact(void);
 
 		const QF* get_cqf(void) const { return &cqf; }
-		uint64_t range(void) const { return cqf.metadata->range; }
+		__uint128_t range(void) const { return cqf.metadata->range; }
 		uint32_t seed(void) const { return cqf.metadata->seed; }
 		uint64_t numslots(void) const { return cqf.metadata->nslots; }
 		uint32_t keybits(void) const { return cqf.metadata->key_bits; }
